Moving the randomPort() function in vttest.
So we can customize it in environments that can reserve ports.
