+++ go/vt/vttest/local_cluster.go
@@ -24,13 +24,11 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"math/rand"
 	"os"
 	"os/exec"
 	"path"
 	"strconv"
 	"strings"
-	"time"
 
 	log "github.com/golang/glog"
 	"github.com/golang/protobuf/proto"
@@ -430,14 +428,4 @@ func (hdl *Handle) run(
 			"error (%v) parsing JSON output from command: %v", err, launcher)
 	}
 	return err
-}
-
-// randomPort returns a random number between 10k & 30k.
-func randomPort() int {
-	v := rand.Int31n(20000)
-	return int(v + 10000)
-}
-
-func init() {
-	rand.Seed(time.Now().UnixNano())
 }
